requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
python-dotenv~=1.0.1